requirements.txt
fastapi==0.104.1
python-restcountries==2.0.0
uvicorn==0.24.0
